docker/AthenaApi/backend/requirements.txt
Flask
flask-cors
pinecone-client
openai
langchain